+++ client/src/components/TokenModal/TokenSelector.jsx
@@ -1,6 +1,5 @@
 import {
 	Flex,
-	Text,
 	Input,
 	Modal,
 	ModalOverlay,
@@ -9,24 +8,19 @@ import {
 	ModalBody,
 	ModalHeader,
 	Divider,
-	Image,
 	InputLeftElement,
 	InputGroup,
 	Heading,
-	Badge,
-	useColorModeValue,
 } from '@chakra-ui/react';
 import { GrFormSearch } from 'react-icons/gr';
+import { SuperFluidContext } from '../../context/SuperFluidContext';
+import TokenCard from './TokenCard';
 // import { screen } from '@testing-library/react';
 
-const TokenSelector = ({
-	isOpen,
-	onClose,
-	tokenList,
-	userTokenList,
-	data,
-	setData,
-}) => {
+const TokenSelector = ({ isOpen, onClose, data, setData }) => {
+	const { defaultTokenList: tokenList, userTokenList } =
+		useContext(SuperFluidContext);
+
 	return (
 		<Modal
 			isOpen={isOpen}
@@ -85,45 +79,6 @@ const TokenSelector = ({
 				</ModalBody>
 			</ModalContent>
 		</Modal>
-	);
-};
-
-const TokenCard = ({ icon, token, symbol, data, setData, onClose }) => {
-	return (
-		<Flex
-			_hover={{
-				bg: useColorModeValue('green.100', 'green.800'),
-				cursor: 'pointer',
-			}}
-			w='100%'
-			align={'center'}
-			justify={'space-between'}
-			borderRadius='md'
-			py={4}
-			px={10}
-			my={1}
-			onClick={() => {
-				setData({ ...data, token: symbol.toUpperCase() });
-				onClose();
-			}}>
-			<Flex align='center'>
-				<Image
-					boxSize='30px'
-					src={icon}
-					alt={symbol}
-					mr={5}
-					borderRaius='md'
-					fallbackSrc={
-						'https://cdn4.iconfinder.com/data/icons/ionicons/512/icon-ios7-circle-outline-512.png'
-					}
-				/>
-				<Text>{token}</Text>
-			</Flex>
-
-			<Badge variant='outline' colorScheme='green' borderRadius='sm'>
-				{symbol}
-			</Badge>
-		</Flex>
 	);
 };
