+++ client/src/pages/CreateStroll.jsx
@@ -1,13 +1,5 @@
-import TokenSelector from '../components/TokenSelector';
-import { useContext, useEffect, useState } from 'react';
-import { DateRange } from 'react-date-range';
-import { addDays } from 'date-fns';
 import {
-	Box,
 	useDisclosure,
-	Flex,
-	FormControl,
-	FormLabel,
 	Button,
 	Modal,
 	ModalOverlay,
@@ -16,37 +8,16 @@ import {
 	ModalFooter,
 	ModalBody,
 	ModalCloseButton,
-	useCheckbox,
-	useCheckboxGroup,
-	Text,
-	useColorModeValue,
 } from '@chakra-ui/react';
-import { SuperFluidContext } from '../context/SuperFluidContext';
-import 'react-date-range/dist/styles.css'; // main css file
-import 'react-date-range/dist/theme/default.css'; // theme css file
+import TokenSelector from '../components/TokenModal/TokenSelector';
+import CreateForm from '../components/CreateForm/CreateForm';
 
 const CreateStroll = ({ isOpen, onClose, data, setData }) => {
-	const { defaultTokenList, userTokenList } = useContext(SuperFluidContext);
-	const { token } = data;
 	const {
 		isOpen: isTokenOpen,
 		onOpen: onTokenOpen,
 		onClose: onTokenClose,
 	} = useDisclosure();
-
-	const { value, getCheckboxProps } = useCheckboxGroup({
-		defaultValue: ['AAVE Tokens'],
-	});
-
-	useEffect(() => setData({ ...data, method: value }), [value]);
-
-	const [dateRange, setDateRange] = useState([
-		{
-			startDate: new Date(),
-			endDate: addDays(new Date(), 7),
-			key: 'selection',
-		},
-	]);
 
 	return (
 		<>
@@ -54,8 +25,6 @@ const CreateStroll = ({ isOpen, onClose, data, setData }) => {
 				isOpen={isTokenOpen}
 				onOpen={onTokenOpen}
 				onClose={onTokenClose}
-				tokenList={defaultTokenList}
-				userTokenList={userTokenList}
 				data={data}
 				setData={setData}
 			/>
@@ -72,69 +41,7 @@ const CreateStroll = ({ isOpen, onClose, data, setData }) => {
 					<ModalHeader>Create a Stroll</ModalHeader>
 					<ModalCloseButton />
 					<ModalBody>
-						<Flex align={'center'} justify={'center'}>
-							<FormControl width='90%'>
-								<FormLabel
-									htmlFor='token'
-									textTransform='uppercase'
-									fontSize='sm'
-									fontWeight='bold'
-									color='gray.300'>
-									Token
-								</FormLabel>
-								<Flex align='center' justify='space-between' mb={7}>
-									<Text
-										py={2}
-										px={8}
-										w='30ch'
-										bg={useColorModeValue('gray.100', 'gray.600')}
-										borderRadius='md'>
-										{token ? token : 'Select a Token'}
-									</Text>
-									<Button onClick={onTokenOpen}>
-										{token ? 'Change Token' : 'Select a Token'}
-									</Button>
-								</Flex>
-								<FormLabel
-									htmlFor='token'
-									textTransform='uppercase'
-									fontSize='sm'
-									fontWeight='bold'
-									color='gray.300'>
-									Investment Methods
-								</FormLabel>
-								<Flex align='center' justify='space-evenly' mb={7}>
-									<CheckboxCard {...getCheckboxProps({ value: 'AAVE Tokens' })}>
-										AAVE Tokens
-									</CheckboxCard>
-									<CheckboxCard
-										isDisabled
-										{...getCheckboxProps({ value: 'ERC20 Tokens' })}>
-										ERC20 Tokens
-									</CheckboxCard>
-								</Flex>
-								<FormLabel
-									htmlFor='duration'
-									textTransform='uppercase'
-									fontSize='sm'
-									fontWeight='bold'
-									color='gray.300'>
-									Duration
-								</FormLabel>
-								<Flex align='center' justify='center' mb={7}>
-									<DateRange
-										editableDateInputs={true}
-										minDate={new Date()}
-										onChange={(item) => {
-											if (item.selection.startDate < Date.now())
-												setDateRange([item.selection]);
-										}}
-										moveRangeOnFirstSelection={false}
-										ranges={dateRange}
-									/>
-								</Flex>
-							</FormControl>
-						</Flex>
+						<CreateForm />
 					</ModalBody>
 					<ModalFooter width={'100%'}>
 						<Button variant='ghost' mr={3} onClick={onClose}>
@@ -155,35 +62,5 @@ const CreateStroll = ({ isOpen, onClose, data, setData }) => {
 		</>
 	);
 };
-
-function CheckboxCard(props) {
-	const { getCheckboxProps, getInputProps } = useCheckbox(props);
-
-	const input = getInputProps();
-	const checkbox = getCheckboxProps();
-	return (
-		<Box as='label'>
-			<input {...input} />
-			<Box
-				{...checkbox}
-				cursor={props.isDisabled ? 'not-allowed' : 'pointer'}
-				borderWidth='1px'
-				borderRadius='md'
-				boxShadow='md'
-				_checked={{
-					bg: 'green.600',
-					color: 'white',
-					borderColor: 'green.600',
-				}}
-				_focus={{
-					boxShadow: 'outline',
-				}}
-				px={5}
-				py={3}>
-				{props.children}
-			</Box>
-		</Box>
-	);
-}
 
 export default CreateStroll;
